move blog form and news form to views
