+++ packages/client/src/app/pages/my-stuff/views/news-form/news-form.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'news-form',
+    templateUrl: './news-form.component.html',
+    styleUrls: ['./news-form.component.less']
+})
+export class NewsFormComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}
